Fixing the interrupt 2 issue
I updated the code for interrupt2 to now delay for 1000 microseconds. This is because the pin kept on having an issue where it would change every time the switch was pressed. I also moved the pin to be on an analog pin so that it is not triggered by mild voltages.